pyproject.toml
[project]
name = "uwm-llm-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain-groq>=0.3.2",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
